don't use popitem to invalidate an LTI auth
When you've launched too many LTI sessions in one session, we invalidate one in order to avoid storing too much information.
The code was written to assume that you could keep an OrderedDict in request.session, but it seems that that isn't the case, in Django 1.9 at least - it gets turned into a normal dict.
Instead, this adds a 'launch_time' key to each LTI launch, and we invalidate the launch with the earliest launch time.