throw a sensible error message when no user ID is present in LTI launch data
the authentication backend looks in the following fields for a username:
* lis_person_sourcedid
* canvas_user_id
* user_id

If none of those are present, then we can't proceed - we need a way of
uniquely identifying the user.

This commit catches that case, and throws an error message "No
user-identifying information was present in the LTI launch data".

I've seen this happen once in Blackboard, though I suspect that's a
configuration problem.